requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.0.4
colorama==0.4.4
dnspython==2.2.1
email-validator==1.1.3
fastapi==0.75.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
Mako==1.2.0
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
pydantic==1.9.0
python-dotenv==0.19.2
sniffio==1.2.0
SQLAlchemy==1.4.32
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.6
